requirements.txt
certifi @ file:///croot/certifi_1671487769961/work/certifi
nvidia-ml-py3==7.352.0
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
PyYAML @ file:///croot/pyyaml_1670514731622/work
setproctitle @ file:///tmp/build/80754af9/setproctitle_1611517561796/work
